various fixes to make code compilable under solaris
git-svn-id: https://svn.code.sf.net/p/pcapsipdump/code/trunk@14 89f3ae11-3634-4b71-aec2-a73e79b8c089

+++ calltable.cpp
@@ -42,7 +42,7 @@ int calltable::add(
 	unsigned long call_id_len,
 	time_t time)
 {
-    unsigned long idx=-1,i,found_empty=0;
+    uint32_t idx=(uint32_t)-1,i,found_empty=0;
     for (i=0;i<table_size;i++){
 	if(table[i].is_used==0){
 	    idx=i;
